src/errors/__init__.py
from jwt import InvalidTokenError

from .auth import AuthenticationError, AuthorizationError
from .user import UserAlreadyExistError, UserNotFoundError

__all__: list[str] = [
    "AuthenticationError",
    "AuthorizationError",
    "InvalidTokenError",
    "UserAlreadyExistError",
    "UserNotFoundError",
]
